move bill management code from dashboard to bill management component
created bill management and calendar-view components

created time-graph and calendar widgets
