+++ src/app/app-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { PageNotFoundComponent } from './shared/components';
 import { DashboardComponent } from "./components/dashboard/dashboard.component";
+import {BillManagementComponent} from "./components/bill-management/bill-management.component";
 
 const routes: Routes = [
   {
@@ -12,6 +13,10 @@ const routes: Routes = [
   {
     path: 'dashboard',
     component: DashboardComponent,
+  },
+  {
+    path: 'bill-management',
+    component: BillManagementComponent,
   },
   {
     path: '**',
