+++ src/app/components/calendar-view/calendar-view.component.html
@@ -1,0 +1,1 @@
+<p>calendar-view works!</p>
