+++ src/app/components/calendar-view/calendar-view.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-calendar-view',
+  standalone: true,
+  imports: [],
+  templateUrl: './calendar-view.component.html',
+  styleUrl: './calendar-view.component.scss'
+})
+export class CalendarViewComponent {
+
+}
