+++ src/app/widgets/calendar/calendar.component.html
@@ -1,0 +1,1 @@
+<p>calendar works!</p>
