+++ src/app/widgets/calendar/calendar.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-calendar',
+  standalone: true,
+  imports: [],
+  templateUrl: './calendar.component.html',
+  styleUrl: './calendar.component.scss'
+})
+export class CalendarComponent {
+
+}
