+++ src/app/widgets/time-graph/time-graph.component.html
@@ -1,0 +1,1 @@
+<p>time-graph works!</p>
