+++ src/app/widgets/time-graph/time-graph.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-time-graph',
+  standalone: true,
+  imports: [],
+  templateUrl: './time-graph.component.html',
+  styleUrl: './time-graph.component.scss'
+})
+export class TimeGraphComponent {
+
+}
